[FIX] hw_drivers: stale token used in send_all_devices
Since the PR odoo/odoo#218109, the send_all_devices logic was refactored
to always send whenever something changes. However, there was a bug
introduced due to the fact the token is saved at start-up and never
updated. This caused the devices to not get sent correctly immediately
after pairing with a DB (in the time before checking out and restarting).

closes odoo/odoo#221178

Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>