[FIX] pos_sale: correctly compute qty_delivered
If you made a sale with multiple lines having the same product, the
qty_delivered was not correctly computed on the different lines.

Steps to reproduce:
-------------------
* Activate the ship later option in the PoS settings
* Open PoS
* Add any product with a quantity of 5
* Apply a discount of 20% to the line
* Click on the same product so that they are added as a second line
* Change the quantity of the second line to 3
* Validate the order and use the ship later option
* Validate the picking
* Go to the sale report and check the qty_delivered for the product
> Observation: The qty_delivered is doubled, 16 instead of 8.

Why the fix:
------------
The qty_delivered was computed by summing the quantities of all moves
for the product, but it did not take into account that the same product
could be present in multiple lines of the same order. We now make sure
to dispatch the delivered quantity correctly across the different lines
of the order, ensuring that the qty_delivered is accurate.

opw-4826487

closes odoo/odoo#221373

Signed-off-by: David Monnom (moda) <moda@odoo.com>