[FIX] crm: Adapt `Lead/Opportunities` action domain props type
Steps:
- Install `crm`
- Enable dev mode
- Dashboard -> Activities icon -> Lead/Opportunitiess
- Traceback

This is simply due to the fact that the withsearch props indicate that
they want to receive an array of strings or arrays, whereas the crm
patch overrides the action domain by keeping the “Domain” type,
which is an object and not an array.

One solution is to use `.toList()` on the domain to convert it into a 2d
array.

opw-4979173

closes odoo/odoo#221602

X-original-commit: 3e25650a0885e428dc3aa8c2817a58bfd6851e50
Related: odoo/enterprise#91556
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>